requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.1
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncer==0.0.2
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boto3==1.35.11
botocore==1.35.11
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chainlit==1.1.200
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.9.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
crewai==0.41.1
crewai-tools==0.4.26
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
docx2txt==0.8
effdet==0.4.1
embedchain==0.1.121
emoji==2.12.1
entrypoints==0.4
executing==2.1.0
faiss-cpu==1.8.0.post1
Faker==28.1.0
fastapi==0.110.3
fastapi-socketio==0.0.10
fastavro==1.9.5
fastembed==0.3.6
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-aiplatform==1.64.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.0.3
groq==0.11.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
humanize==4.10.0
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
instructor==1.3.3
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
json_repair==0.25.3
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.6
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.16
langchain-cohere==0.1.9
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.63
langchain-groq==0.1.6
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.111
lark==1.1.9
layoutparser==0.3.4
Lazify==0.4.0
literalai==0.0.601
loguru==0.7.2
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.0.20
merkle-json==1.0.0
millify==0.1.1
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
munch==2.5.0
mypy-extensions==1.0.0
narwhals==1.6.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemoguardrails==0.9.1.1
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.0
openai==1.43.0
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pickleshare==0.7.5
pikepdf==9.2.1
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.24.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.6.3
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.8
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
pymdown-extensions==10.9
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.378
pysbd==0.3.4
PySocks==1.7.1
PyStemmer==2.2.0.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
python-oxmsg==0.0.1
python-socketio==5.11.4
pytube==15.0.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.11.1
rapidfuzz==3.9.7
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.24.0
semantic-router==0.0.63
semver==3.0.2
Send2Trash==1.8.3
setuptools==73.0.1
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.0.0
simpleeval==0.9.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.33
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.2
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
types-python-dateutil==2.9.0.20240821
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.9
unstructured-client==0.18.0
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
uptrace==1.26.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.25.0
validators==0.34.0
watchdog==4.0.2
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
wheel==0.44.0
widgetsnbextension==4.0.13
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.8
zipp==3.20.1
trulens==1.0.3
trulens-apps-langchain==1.0.3
trulens-providers-openai==1.0.3
